backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-JWT-Extended
Flask-Bcrypt
Flask-Migrate
python-dotenv
